+++ fusion/skins/rtai/cond.c
@@ -47,12 +47,12 @@
 
 #ifdef CONFIG_RTAI_NATIVE_EXPORT_REGISTRY
 
-static ssize_t __cond_read_proc (char *page,
-				 char **start,
-				 off_t off,
-				 int count,
-				 int *eof,
-				 void *data)
+static int __cond_read_proc (char *page,
+			     char **start,
+			     off_t off,
+			     int count,
+			     int *eof,
+			     void *data)
 {
     RT_COND *cond = (RT_COND *)data;
     char *p = page;
