+++ fusion/skins/rtai/sem.c
@@ -47,12 +47,12 @@
 
 #ifdef CONFIG_RTAI_NATIVE_EXPORT_REGISTRY
 
-static ssize_t __sem_read_proc (char *page,
-				char **start,
-				off_t off,
-				int count,
-				int *eof,
-				void *data)
+static int __sem_read_proc (char *page,
+			    char **start,
+			    off_t off,
+			    int count,
+			    int *eof,
+			    void *data)
 {
     RT_SEM *sem = (RT_SEM *)data;
     char *p = page;
